Added part 3: Rainbow Table Attack
Generates a Rainbow Table as rainbowTable.txt given input file english3.txt
Passwords can be cracked by copying hexcode from the password file and using "ctrl+f" to find matching entries in rainbowTable.txt.  This is working.

A better implementation can be parsing the rainbow table within the terminal window, if we have time for that.